Feat: 복주머니 조회, 복주머니 주소 조회

+++ back/luckybocky/src/main/java/com/project/luckybocky/fortune/entity/Fortune.java
@@ -13,8 +13,8 @@ import lombok.*;
 public class Fortune extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name ="fortune_seq")
-    private Integer fortuneSeq;
+    @Column(name = "fortune_seq")
+    private Short fortuneSeq;
 
     @Column(length = 8, nullable = false)
     private String fortuneName;
